updates to requirements
