+++ lib/models/artistdb.py
@@ -1,12 +1,13 @@
 import sqlalchemy
-from sqlalchemy import Table, Column, Integer, Float, ForeignKey, MetaData
+from sqlalchemy import Table, Column, Integer, Float, String, ForeignKey, MetaData
 from sqlalchemy.orm import relationship
 from lib.models.declarative_base import DeclarativeBase
 
-
-class YearDB(DeclarativeBase):
-    __tablename__ = 'yeardb'
-    release_year = Column(Integer, primary_key=True)
+class ArtistDB(DeclarativeBase):
+    __tablename__ = 'artistdb'
+    artists_id = Column(String(100), primary_key=True)
+    artists = Column(String(100))
+    artist_popularity = Column(Integer)
     acousticness_mean = Column(Float)
     danceability_mean = Column(Float)
     duration_ms_mean = Column(Float)
@@ -18,11 +19,15 @@ class YearDB(DeclarativeBase):
     tempo_mean = Column(Float)
     valence_mean = Column(Float)
     popularity_mean = Column(Float)
-    music_key_mode = Column(Integer, ForeignKey('musickey.music_key'))
+    music_key_mode = Column(Integer, ForeignKey('musickeydb.music_key'))
     major_minor_mode = Column(Integer)
+    num_tracks = Column(Integer)
+    genre = Column(String(100))
+    followers = Column(Integer, default=0)
+    musickey = relationship("MusicKeyDB")
 
     def __init__(self):
-        self.name = 'yeardb'
+        self.name = 'test'
         #self.meta = MetaData()
 
     def create_table(self,connection):
