+++ lib/models/genredb.py
@@ -3,11 +3,9 @@ from sqlalchemy import Table, Column, Integer, Float, String, ForeignKey, MetaDa
 from sqlalchemy.orm import relationship
 from lib.models.declarative_base import DeclarativeBase
 
-class ArtistDB(DeclarativeBase):
-    __tablename__ = 'artistdb'
-    artists_id = Column(String(100), primary_key=True)
-    artists = Column(String(100))
-    artist_popularity = Column(Integer)
+class GenreDB(DeclarativeBase):
+    __tablename__ = 'genredb'
+    genres = Column(String(100), primary_key=True)
     acousticness_mean = Column(Float)
     danceability_mean = Column(Float)
     duration_ms_mean = Column(Float)
@@ -19,14 +17,12 @@ class ArtistDB(DeclarativeBase):
     tempo_mean = Column(Float)
     valence_mean = Column(Float)
     popularity_mean = Column(Float)
-    music_key_mode = Column(Integer, ForeignKey('musickey.music_key'))
+    music_key_mode = Column(Integer, ForeignKey('musickeydb.music_key'))
     major_minor_mode = Column(Integer)
-    num_tracks = Column(Integer)
-    genre = Column(String(100))
-    followers = Column(Integer, default=0)
+    musickey = relationship("MusicKeyDB")
 
     def __init__(self):
-        self.name = 'test'
+        self.name = 'genredb'
         #self.meta = MetaData()
 
     def create_table(self,connection):
@@ -34,4 +30,4 @@ class ArtistDB(DeclarativeBase):
         if not engine.dialect.has_table(engine, self.name):
             self.table.create(engine)
         else:
-            print(self.name, 'database already exists')
+            print(self.name,'database already exists')
