+++ requirements/dev.txt
@@ -1,0 +1,1 @@
+-r prod.txt
