--- requirements/prod
@@ -1,2 +1,0 @@
-mysql-connector-python = 8.0.20
-sqlalchemy = 1.3.16
