+++ requirements/prod.txt
@@ -1,0 +1,5 @@
+mysql-connector-python == 8.0.20
+sqlalchemy == 1.3.16
+marshmallow_sqlalchemy == 0.24.2
+pymysql == 1.0.2
+pandas
